Move generate_bench_website to utils
